refactor to be a mixin instead
